+++ lib/delta_trace_db.dart
@@ -22,6 +22,7 @@ export 'src/query/enum_query_type.dart';
 export 'src/query/query.dart';
 export 'src/query/query_builder.dart';
 export 'src/query/query_result.dart';
+export 'src/query/raw_query_builder.dart';
 
 // internal only
 // export 'src/query/util_filed.dart';